--- RecoLocalTracker/SiPixelClusterizer/BuildFile.xml
@@ -1,8 +1,0 @@
-<use   name="DataFormats/Common"/>
-<use   name="FWCore/ParameterSet"/>
-<use   name="DataFormats/SiPixelDetId"/>
-<use   name="DataFormats/SiPixelCluster"/>
-<use   name="boost_serialization"/>
-<export>
-  <lib   name="1"/>
-</export>
