+++ RecoLocalTracker/SiPixelClusterizer/plugins/BuildFile.xml
@@ -1,4 +1,8 @@
-<use   name="RecoLocalTracker/SiPixelClusterizer"/>
+<use   name="DataFormats/Common"/>
+<use   name="FWCore/ParameterSet"/>
+<use   name="DataFormats/SiPixelDetId"/>
+<use   name="DataFormats/SiPixelCluster"/>
+<use   name="boost_serialization"/>
 <use   name="CalibTracker/SiPixelESProducers"/>
 <library   file="*.cc" name="RecoLocalTrackerSiPixelClusterizerPlugins">
   <flags   EDM_PLUGIN="1"/>
