+++ RecoLocalTracker/SiPixelClusterizer/plugins/SiPixelClusterProducer.cc
@@ -12,8 +12,8 @@
  */
 
 // Our own stuff
-#include "RecoLocalTracker/SiPixelClusterizer/interface/SiPixelClusterProducer.h"
-#include "RecoLocalTracker/SiPixelClusterizer/interface/PixelThresholdClusterizer.h"
+#include "SiPixelClusterProducer.h"
+#include "PixelThresholdClusterizer.h"
 
 // Geometry
 #include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
@@ -42,8 +42,6 @@
 // MessageLogger
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 
-namespace cms
-{
 
   //---------------------------------------------------------------------------
   //!  Constructor: set the ParameterSet and defer all thinking to setupClusterizer().
@@ -83,12 +81,6 @@ namespace cms
     delete theSiPixelGainCalibration_;
   }  
 
-  //void SiPixelClusterProducer::beginJob( const edm::EventSetup& es ) 
-  void SiPixelClusterProducer::beginJob( ) 
-  {
-    edm::LogInfo("SiPixelClusterizer") << "[SiPixelClusterizer::beginJob]";
-    clusterizer_->setSiPixelGainCalibrationService(theSiPixelGainCalibration_);
-  }
   
   //---------------------------------------------------------------------------
   //! The "Event" entrypoint: gets called by framework for every event
@@ -132,6 +124,7 @@ namespace cms
 
     if ( clusterMode_ == "PixelThresholdClusterizer" ) {
       clusterizer_ = new PixelThresholdClusterizer(conf_);
+      clusterizer_->setSiPixelGainCalibrationService(theSiPixelGainCalibration_);
       readyToCluster_ = true;
     } 
     else {
@@ -204,4 +197,11 @@ namespace cms
     //				    << " SiPixelClusters in " << numberOfDetUnits << " DetUnits."; 
   }
 
-}  // end of namespace cms
+
+
+
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+
+DEFINE_FWK_MODULE(SiPixelClusterProducer);
+
